Put all ariadne wrapping in diagnostic module
