+++ crates/apollo-compiler/src/lib.rs
@@ -4,7 +4,7 @@
 mod macros;
 pub mod ast;
 mod database;
-mod diagnostic;
+pub mod diagnostic;
 mod diagnostics;
 pub mod executable;
 pub mod execution;
